chore: Rename executor.ts to interpreter.ts
